backport libdave fixes from discord https://github.com/discord/libdave/commit/6e5ffbc1cb4eef6be96e8115c4626be598b7e501#diff-f395bfd18d0d8d925834eef8a2732beff51c7bab0cb55b2901eacf3c01a39899R133

+++ src/dpp/dave/frame_processors.h
@@ -73,7 +73,7 @@ uint8_t serialize_unencrypted_ranges(const ranges& unencrypted_ranges, uint8_t* 
  * @param unencrypted_ranges unencrypted ranges to write to
  * @return size of unencrypted ranges written
  */
-uint8_t deserialize_unencrypted_ranges(const uint8_t*& read_at, const size_t buffer_size, ranges& unencrypted_ranges);
+uint8_t deserialize_unencrypted_ranges(const uint8_t*& read_at, const uint8_t buffer_size, ranges& unencrypted_ranges);
 
 /**
  * @brief Validate unencrypted ranges
